--- a/test/srt/test_vision_openai_server_a.py
+++ b/test/srt/test_vision_openai_server_a.py
@@ -31,6 +31,10 @@ class TestQwen2VLServer(TestOpenAIVisionServer):
             other_args=[
                 "--mem-fraction-static",
                 "0.35",
+                "--mm-attention-backend",
+                "fa3",
+                "--cuda-graph-max-bs",
+                "4",
             ],
         )
         cls.base_url += "/v1"
@@ -53,6 +57,10 @@ class TestQwen2_5_VLServer(TestOpenAIVisionServer):
             other_args=[
                 "--mem-fraction-static",
                 "0.35",
+                "--mm-attention-backend",
+                "fa3",
+                "--cuda-graph-max-bs",
+                "4",
             ],
         )
         cls.base_url += "/v1"
@@ -76,6 +84,10 @@ class TestVLMContextLengthIssue(CustomTestCase):
                 "--context-length",
                 "300",
                 "--mem-fraction-static=0.75",
+                "--mm-attention-backend",
+                "fa3",
+                "--cuda-graph-max-bs",
+                "4",
             ],
         )
         cls.base_url += "/v1"
@@ -149,6 +161,10 @@ class TestMinicpmvServer(TestOpenAIVisionServer):
                 "--trust-remote-code",
                 "--mem-fraction-static",
                 "0.35",
+                "--mm-attention-backend",
+                "fa3",
+                "--cuda-graph-max-bs",
+                "4",
             ],
         )
         cls.base_url += "/v1"
@@ -164,7 +180,13 @@ class TestInternVL2_5Server(TestOpenAIVisionServer):
             cls.model,
             cls.base_url,
             timeout=DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
-            other_args=["--trust-remote-code"],
+            other_args=[
+                "--trust-remote-code",
+                "--mm-attention-backend",
+                "fa3",
+                "--cuda-graph-max-bs",
+                "4",
+            ],
         )
         cls.base_url += "/v1"
 
@@ -183,6 +205,10 @@ class TestMinicpmoServer(TestOpenAIVisionServer):
                 "--trust-remote-code",
                 "--mem-fraction-static",
                 "0.65",
+                "--mm-attention-backend",
+                "fa3",
+                "--cuda-graph-max-bs",
+                "4",
             ],
         )
         cls.base_url += "/v1"
@@ -207,10 +233,15 @@ class TestMimoVLServer(TestOpenAIVisionServer):
                 "--trust-remote-code",
                 "--mem-fraction-static",
                 "0.6",
+                "--mm-attention-backend",
+                "fa3",
+                "--cuda-graph-max-bs",
+                "4",
             ],
         )
         cls.base_url += "/v1"
 
 
 if __name__ == "__main__":
+    del TestOpenAIVisionServer
     unittest.main()