<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation
TODO
Add coverage to tp-2/4


## Modifications

- Enhance test configurations by adding new arguments for mm-attention-backend and cuda-graph-max-bs across multiple vision server tests.


```
Before modify:
python -m unittest test_vision_openai_server_a  243.56s user 32.62s system 49% cpu 9:22.73 total
python -m unittest test_vision_openai_server_a  244.91s user 34.20s system 49% cpu 9:24.91 total


After modify:
python -m unittest test_vision_openai_server_a  247.31s user 31.40s system 64% cpu 7:12.95 total
python -m unittest test_vision_openai_server_a  247.28s user 34.04s system 65% cpu 7:12.18 total
```
<!-- Describe the changes made in this PR. -->

## Checklist

- [ ] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [ ] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.